chore: infer when possible, move schemas, add tsconfig to scripts
