--- packages/mcp-server/src/mcp/schemas/index.ts
@@ -1,19 +1,0 @@
-import * as v from 'valibot';
-
-export const documentation_sections_schema = v.record(
-	v.string(),
-	v.object({
-		metadata: v.object({
-			title: v.string(),
-			use_cases: v.optional(v.string()),
-		}),
-		slug: v.string(),
-	}),
-);
-
-export const section_schema = v.object({
-	title: v.string(),
-	use_cases: v.string(),
-	slug: v.string(),
-	url: v.string(),
-});
